+++ content_editor/__init__.py
@@ -1,2 +1,2 @@
-VERSION = (6, 2, 0)
+VERSION = (6, 2, 1)
 __version__ = ".".join(map(str, VERSION))